framegrabber: Move UMat alloc to constructor
Move allocation of UMats from the get function to the constructor of the
class. This makes the code simpler but assumes that UMat will preserve the
pointer reference so that it can be updated...

A nice bonus is that height and width no longer needs to be saved in the
class as members.